Update commandline-testing_version.py
I have added the ability to add a task as well as a sub-task. Furthermore, I added the ability to view the JSON file directly with a print statement in a raw form for testing, and this function can also be used in other functions in order for them to access the JSON file.