Update commandline-testing_version.py
I made the code look a bit cleaner in addition to that I added the ability to view the tasks in a "table". Furthermore, I added the ability to select a task and add a sub-task to it. Additionally, I  added the ability to exit the program and rename a task.