Update commandline-testing_version.py
Added the ability to delete and change the status of a task or a sub-task. if the task has sub-tasks it will prompt the user with 2 choices(if the user wants to apply the action to the task or the sub-task). If it doesn't it will auto apply the action to the task.